% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.returnOne.R
\name{ds.returnOne}
\alias{ds.returnOne}
\title{Title}
\usage{
ds.returnOne(x = NULL, datasources = NULL)
}
\arguments{
\item{datasources}{}
}
\value{
This function simply returns the integer value 1
}
\description{
Title
}
\examples{

}
